Remove prefix settings for danger event
This is to fix a bug in the message ID. DangerMovement can't know the
prefix used in the DangerMessageGenerator class. A fix for keeping the
prefix settings would be to set it for both the generator and the
model movement. Current fix is to set all danger message with the
prefix "DANGER".